Feature/sxl 1.1.2

+++ lib/cocoapods-meitu-bin/helpers/buildAll/builder.rb
@@ -1,3 +1,4 @@
+require 'cocoapods-meitu-bin/config/config'
 
 module CBin
   module BuildAll
@@ -11,6 +12,9 @@ module CBin
         @configuration = configuration
         @file_accessors = pod_target.file_accessors unless pod_target.nil?
         @base_dir = "#{Pathname.pwd}/build_pods"
+        if PodUpdateConfig.shell_project
+          @base_dir = "#{Pathname.pwd}/all_build/Build"
+        end
       end
 
       # 构建
@@ -92,7 +96,7 @@ module CBin
 
       # xxx.framework 所在目录
       def product_dir
-        @product_dir = "#{@base_dir}/#{@pod_target}/Products"
+        @product_dir = "#{@base_dir}/Products"
         @product_dir
       end
 
@@ -348,8 +352,22 @@ ibtool \
         fwks_dir = "#{result_product_dir}/fwks"
         FileUtils.mkdir(fwks_dir) unless File.exist?(fwks_dir)
         fwks.map do |fwk|
-          `rsync -av #{fwk} #{fwks_dir}`
+          new_fwk = get_XCFrameworkIntermediates_path(fwk)
+          if new_fwk && File.exist?(new_fwk)
+            `rsync -av #{new_fwk} #{fwks_dir}`
+          else
+            `rsync -av #{fwk} #{fwks_dir}`
+          end
         end
+      end
+      def get_XCFrameworkIntermediates_path(path)
+        # 使用正则表达式匹配并提取 AdsFramework
+        match = path.to_s.match(/\/([^\/]+)\.xcframework/)
+        if match
+          framework_path = "#{product_dir}/#{iphoneos}/XCFrameworkIntermediates/#{match[1]}/#{match[1]}.framework"
+          return framework_path
+        end
+        return  nil
       end
 
       # 拷贝 framework
